+++ src/Utils.c
@@ -7,10 +7,10 @@
 #define MIN(a, b) ((a < b) ? a : b)
 #define MAX(a, b) ((a > b) ? a : b)
 
-static float _angle(float ax, float ay, float bx, float by) {
+static double _angle(double ax, double ay, double bx, double by) {
     return atan2(bx - ax, by - ay);
 }
 
-static float _degrees(float radians) {
+static double _degrees(double radians) {
     return remainder(radians * (180.0 / M_PI), 360.0);
 }
